deployer: Improve output target generation
This patch forces output target generation to occur when the resolve()
callback is called for the Target's exec() Promise. This forces a
connection between the set of returned Targets and those added to the
output list. Previously there was no such connection, which could have
allowed for inconsistency between the output list and the list of logged
target interactions that the user would see on the console.