Improve build plugin run-mode behavior
This patch enables empty plugin handler lists to be used to imply
passing targets generated by an include section straight to the deploy
plugin. It allows for an include section to be ignored entirely if the
run-mode is build. This improves the flexibility of the config schema.

Furthermore, this patch made the build flag default to true for any
handler.