[combine] Fix handling of new targets that have build handlers
This patch fixes an issue where an output target got placed in both the
initial target set and the output target set simultaneously. This
happened when the target was pushed to process its build handlers while
also being added to the output set. These changes depend on recent
changes to the DeployContext class.