[builder] Add initial target loading from tree
This commit improves the Builder's initial target loading methods and
allows for loading directly from the target tree (without checking if
the target matches an include).

The generic push method (i.e. pushInitialTarget()) was simplified to
only support Target objects while the pushInitialTargetDelayed() method
was added to do the same thing for delayed targets. The delayed target
handler marks the targets as initial so that the builder can remember
the initial set of targets that were specified to build.