builder: Fix bug opening every traversed file
This one is pretty embarrassing... The builder was loading all delayed
targets before checking if they were included. In this way, the nodejs
process was opening file descriptors for a read stream for every file
traversed for the build.

The fix properly handles delayed targets so that they are only opened if
they are going to be added to the builder.