Improve depend graph functionality with null products
Adds a new feature to the dependency graph that allows a null product
node to exist in the graph. The null product node denotes sources in the
dependency tree that are not a part of the build but *may* be included
at a later time. Such sources are ignored in future builds if no other
source is out-of-date.

The null build product is really designed to prevent running a build due
to sources existing in the include set but not being used by the deploy
plugin. It is up to the deploy plugin to add null connections to the
dependency graph when it identifies a target that is to be excluded from
the build.

This commit also updates the internal structure of DependencyGraph to
use Map for most dictionary purposes. (We had to do this to support an
unambiguous 'null' key.)