Normalize tree path when loading tree record
This patch normalizes the path used to load a tree record. The
normalization results in a path that is the same for both PathTree and
RepoTree instances. This greatly improves usability when a project tree
is used in both a build and deploy context. It also fixes a bug that was
causing default 'targetTree' parameter to not be configured for
'PathTree' instances.